Change type of tls skipVerify from bool to string while parsing secret data
The stringData of the s3 secret cannot accept insecureSkipVerify input as a boolean flag.
The helm chart will now quote that field converting the input from bool to string.
And the backup-restore-operator will unmarshal secret.StringData into a temporary objectStore
struct with insecureSkipVerify as a string, and then set it as a bool in the original s3ObjectStore type