backport of https://github.com/rancher/backup-restore-operator/pull/149 drop the file statussubresource.json and implement a new approach to restore resources the subresource Status
Bug:
While going through the collected resources, the current implementation checks if its name is xxx/status and supports the update operation; if so, add its parent resource kind to a list that is saved to a file. And the list is used as the source of truth for determining if the status of a resource needs to be updated.
And its problem is that if the rules in the resourceSet are strict, we will miss the status resource, therefore not update the status after restoring.

Fix:
The new approach: when restoring the CRDs, if we see the spec.subresources.status filed, it means this kind of resource has the status subresource that needs to be restored. We save the names of those CRD into a list, then when restoring other resources, we will update its status if its kind is in the list.
The new approach is more robust as it checks the CRD directly instead of depending on collected resources.