Force sync fleet-agent bundle to remove modified states
We need to force sync fleet-agent bundle after a restore to prevent
fleet cluster falls into modified state, as there is previously a bug
with fleet that we use two different setID to track fleet-agent bundle.
The bug is fixed in fleet v0.3.6 and above, but in order for fleet
v0.3.5 to work with backup operator we need to add this fix in.