Add PSA configuration secret to provisioning resource set
When a k3s/RKE2 cluster is created with a non-default PSA configuration
template, the configuration from that template is then stored in a secret,
which is used by the control plane planner.

This secret is only updated by the webhook when the provisioning cluster
object is updated, and it wasn't included in the backups created from the
default resource set, so after a migration the planner would get stuck in an
error state until the secret was re-created manually.