refactor: Correctly process files for optionalResources.kubewarden-user-crs.enabled
files under `files/` arent' templates under `templates/`, hence their
go templating doesn't get evaluated.

Move the kubewarden user CRs to a separate file, and add a conditional
to contatenate it to the resourceset templates when the new Value is
set.

This is simpler than refactoring all the `files/` into `templates/`.

Signed-off-by: Víctor Cuadrado Juan <vcuadradojuan@suse.de>